Move ActionButton to own component
